requirements.txt
PyYAML
python-dotenv
requests
FastAPI
uvicorn
APScheduler
ollama
python-multipart
email-validator
imaplib2
SQLAlchemy
beautifulsoup4
python-telegram-bot
discord.py
tiktoken
typing_extensions>=4.6.0
psycopg2-binary==2.9.9
pgvector
praw
slack_sdk
wikipedia-api
google-api-python-client
